feat: command for delete client

+++ src/EcoPark.Application/Employees/Delete/DeleteEmployeeCommandHandler.cs
@@ -8,22 +8,31 @@ public class DeleteEmployeeCommandHandler(IRepository<EmployeeModel> repository)
 
         try
         {
-            await repository.UnitOfWork.StartAsync(cancellationToken);
-
-            var databaseOperationResult = await repository.DeleteAsync(command, cancellationToken);
-
-            if (databaseOperationResult)
+            if (await repository.CheckChangePermissionAsync(command, cancellationToken))
             {
-                await repository.UnitOfWork.SaveEntitiesAsync(cancellationToken);
-                await repository.UnitOfWork.CommitAsync(cancellationToken);
+                await repository.UnitOfWork.StartAsync(cancellationToken);
 
-                result = new DatabaseOperationResponseViewModel("Delete", EOperationStatus.Successful, "Employee was deleted successfully!");
+                var databaseOperationResult = await repository.DeleteAsync(command, cancellationToken);
+
+                if (databaseOperationResult)
+                {
+                    await repository.UnitOfWork.SaveEntitiesAsync(cancellationToken);
+                    await repository.UnitOfWork.CommitAsync(cancellationToken);
+
+                    result = new DatabaseOperationResponseViewModel("Delete", EOperationStatus.Successful,
+                        "Employee was deleted successfully!");
+                }
+                else
+                {
+                    await repository.UnitOfWork.RollbackAsync(cancellationToken);
+                    result = new DatabaseOperationResponseViewModel("Delete", EOperationStatus.Failed,
+                        "No Employee were found with this id");
+                }
             }
             else
-            {
-                await repository.UnitOfWork.RollbackAsync(cancellationToken);
-                result = new DatabaseOperationResponseViewModel("Delete", EOperationStatus.Failed, "No Employee were found with this id");
-            }
+                result = new DatabaseOperationResponseViewModel("Delete", EOperationStatus.NotAuthorized,
+                    "You have no permission to delete this employee");
+
         }
         catch (Exception e)
         {
